faster-backend/pyproject.toml
[project]
name = "faster-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "psycopg2>=2.9.10",
    "whitenoise>=6.9.0",
]
